# pipeline/api-specs/openapi.yaml
components:
  schemas:
...
    AgentArtifactListRequest:
      properties:
        run_id:
          minLength: 1
          title: Run Id
          type: string
        user_id:
          minLength: 1
          title: User Id
          type: string
        workspace_id:
          minLength: 1
          title: Workspace Id
          type: string
      required:
      - run_id
      - workspace_id
      - user_id
      title: AgentArtifactListRequest
      type: object
    AgentArtifactRegisterRequest:
      properties:
        artifact_id:
          minLength: 1
          title: Artifact Id
          type: string
        base_version:
          anyOf:
          - minimum: 1.0
            type: integer
          - type: 'null'
          title: Base Version
        content_hash:
          maxLength: 71
          minLength: 71
          title: Content Hash
          type: string
        document_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Id
        markdown:
          maxLength: 5242880
          minLength: 0
          title: Markdown
          type: string
        purpose:
          minLength: 1
          title: Purpose
          type: string
        run_id:
          minLength: 1
          title: Run Id
          type: string
        target:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Target
        user_id:
          minLength: 1
          title: User Id
          type: string
        workspace_id:
          minLength: 1
          title: Workspace Id
          type: string
      required:
      - run_id
      - workspace_id
      - user_id
      - artifact_id
      - content_hash
      - purpose
      - markdown
      title: AgentArtifactRegisterRequest
      type: object
    AgentArtifactResolveRequest:
      properties:
        artifact_id:
          minLength: 1
          title: Artifact Id
          type: string
        base_version:
          anyOf:
          - minimum: 1.0
            type: integer
          - type: 'null'
          title: Base Version
        content_hash:
          maxLength: 71
          minLength: 71
          title: Content Hash
          type: string
        document_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Id
        purpose:
          minLength: 1
          title: Purpose
          type: string
        run_id:
          minLength: 1
          title: Run Id
          type: string
        target:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Target
        user_id:
          minLength: 1
          title: User Id
          type: string
        workspace_id:
          minLength: 1
          title: Workspace Id
          type: string
      required:
      - run_id
      - workspace_id
      - user_id
      - artifact_id
      - content_hash
      - purpose
      title: AgentArtifactResolveRequest
      type: object
    AgentArtifactResolveResponse:
      properties:
        base_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Base Version
        content_hash:
          title: Content Hash
          type: string
        document_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Id
        id:
          title: Id
          type: string
        markdown:
          title: Markdown
          type: string
        purpose:
          title: Purpose
          type: string
        target:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Target
      required:
      - id
      - content_hash
      - purpose
      - markdown
      title: AgentArtifactResolveResponse
      type: object
    AgentArtifactResponse:
      properties:
        base_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Base Version
        content_hash:
          title: Content Hash
          type: string
        document_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Id
        id:
          title: Id
          type: string
        purpose:
          title: Purpose
          type: string
        target:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Target
      required:
      - id
      - content_hash
      - purpose
      title: AgentArtifactResponse
      type: object
...
    AgentPlanOperationResponse:
      properties:
        arguments:
          additionalProperties: true
          title: Arguments
          type: object
        base_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Base Version
        depends_on:
          items:
            type: string
          title: Depends On
          type: array
        destination_parent_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Destination Parent Id
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
        id:
          title: Id
          type: string
        reason:
          title: Reason
          type: string
        sequence:
          title: Sequence
          type: integer
        source_parent_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Parent Id
        status:
          title: Status
          type: string
        target_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Id
        target_type:
          title: Target Type
          type: string
        tool_name:
          title: Tool Name
          type: string
      required:
      - id
      - sequence
      - tool_name
      - target_type
      - target_id
      - base_version
      - source_parent_id
      - destination_parent_id
      - arguments
      - reason
      - depends_on
      - status
      - error_code
      title: AgentPlanOperationResponse
      type: object
    AgentPlanResponse:
      properties:
        id:
          title: Id
          type: string
        operation_hash:
          title: Operation Hash
          type: string
        operations:
          items:
            $ref: '#/components/schemas/AgentPlanOperationResponse'
          title: Operations
          type: array
        status:
          title: Status
          type: string
        summary:
          title: Summary
          type: string
        version:
          title: Version
          type: integer
      required:
      - id
      - version
      - summary
      - operation_hash
      - status
      - operations
      title: AgentPlanResponse
      type: object
    AgentRunActorRequest:
      properties:
        user_id:
          minLength: 1
          title: User Id
          type: string
        workspace_id:
          minLength: 1
          title: Workspace Id
          type: string
      required:
      - workspace_id
      - user_id
      title: AgentRunActorRequest
      type: object
    AgentRunResponse:
      properties:
        action:
          title: Action
          type: string
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
        id:
          title: Id
          type: string
        plan:
          anyOf:
          - $ref: '#/components/schemas/AgentPlanResponse'
          - type: 'null'
        request_summary:
          title: Request Summary
          type: string
        skill_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Skill Version Id
        status:
          title: Status
          type: string
        workspace_id:
          title: Workspace Id
          type: string
      required:
      - id
      - workspace_id
      - action
      - skill_version_id
      - status
      - request_summary
      - error_code
      title: AgentRunResponse
      type: object
    AgentToolExecuteAuthorizationRequest:
      properties:
        arguments:
          additionalProperties: true
          title: Arguments
          type: object
        operation_hash:
          maxLength: 64
          minLength: 64
          title: Operation Hash
          type: string
        operation_id:
          minLength: 1
          title: Operation Id
          type: string
        plan_id:
          minLength: 1
          title: Plan Id
          type: string
        plan_version:
          minimum: 1.0
          title: Plan Version
          type: integer
        run_id:
          minLength: 1
          title: Run Id
          type: string
        tool_name:
          minLength: 1
          title: Tool Name
          type: string
        user_id:
          minLength: 1
          title: User Id
          type: string
        workspace_id:
          minLength: 1
          title: Workspace Id
          type: string
      required:
      - workspace_id
      - user_id
      - run_id
      - plan_id
      - plan_version
      - operation_hash
      - operation_id
      - tool_name
      - arguments
      title: AgentToolExecuteAuthorizationRequest
      type: object
    AgentToolReadAuthorizationRequest:
      properties:
        run_id:
          minLength: 1
          title: Run Id
          type: string
        user_id:
          minLength: 1
          title: User Id
          type: string
        workspace_id:
          minLength: 1
          title: Workspace Id
          type: string
      required:
      - workspace_id
      - user_id
      - run_id
      title: AgentToolReadAuthorizationRequest
      type: object
...
    ApproveAgentPlanRequest:
      properties:
        operation_hash:
          maxLength: 64
          minLength: 64
          title: Operation Hash
          type: string
        plan_version:
          minimum: 1.0
          title: Plan Version
          type: integer
        user_id:
          minLength: 1
          title: User Id
          type: string
        workspace_id:
          minLength: 1
          title: Workspace Id
          type: string
      required:
      - workspace_id
      - user_id
      - plan_version
      - operation_hash
      title: ApproveAgentPlanRequest
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    MarkdownAgentRunStatusResponse:
      properties:
        apply_operation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Apply Operation Id
        base_version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Base Version
        document_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Document Id
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
        id:
          title: Id
          type: string
        result:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Result
        status:
          title: Status
          type: string
      required:
      - id
      - status
      - result
      - error_code
      title: MarkdownAgentRunStatusResponse
      type: object
...
    ReviseAgentPlanRequest:
      properties:
        instruction:
          maxLength: 1000
          minLength: 1
          title: Instruction
          type: string
        user_id:
          minLength: 1
          title: User Id
          type: string
        workspace_id:
          minLength: 1
          title: Workspace Id
          type: string
      required:
      - workspace_id
      - user_id
      - instruction
      title: ReviseAgentPlanRequest
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  title: Fruition Pipeline Lab API
  version: 0.1.0
openapi: 3.1.0
paths:
  /agent/runs/{run_id}:
    get:
      operationId: get_agent_run_agent_runs__run_id__get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: query
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      - in: query
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      - in: header
        name: X-Agent-Service-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Service-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Agent Run
      tags:
      - agent-runs
  /agent/runs/{run_id}/approve:
    post:
      operationId: approve_agent_run_agent_runs__run_id__approve_post
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: header
        name: X-Agent-Service-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Service-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveAgentPlanRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Approve Agent Run
      tags:
      - agent-runs
  /agent/runs/{run_id}/cancel:
    post:
      operationId: cancel_agent_run_agent_runs__run_id__cancel_post
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: header
        name: X-Agent-Service-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Service-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRunActorRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Cancel Agent Run
      tags:
      - agent-runs
  /agent/runs/{run_id}/reject:
    post:
      operationId: reject_agent_run_agent_runs__run_id__reject_post
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: header
        name: X-Agent-Service-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Service-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRunActorRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Reject Agent Run
      tags:
      - agent-runs
  /agent/runs/{run_id}/revise:
    post:
      operationId: revise_agent_run_agent_runs__run_id__revise_post
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: header
        name: X-Agent-Service-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Service-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviseAgentPlanRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Revise Agent Run
      tags:
      - agent-runs
...
  /internal/agent/runs/artifacts/list:
    post:
      operationId: list_agent_artifacts_internal_agent_runs_artifacts_list_post
      parameters:
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentArtifactListRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AgentArtifactResponse'
                title: Response List Agent Artifacts Internal Agent Runs Artifacts List Post
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Agent Artifacts
      tags:
      - internal-agent-runs
  /internal/agent/runs/artifacts/register:
    post:
      operationId: register_agent_artifact_internal_agent_runs_artifacts_register_post
      parameters:
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentArtifactRegisterRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentArtifactResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Register Agent Artifact
      tags:
      - internal-agent-runs
  /internal/agent/runs/artifacts/resolve:
    post:
      operationId: resolve_agent_artifact_internal_agent_runs_artifacts_resolve_post
      parameters:
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentArtifactResolveRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentArtifactResolveResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Resolve Agent Artifact
      tags:
      - internal-agent-runs
  /internal/agent/runs/tool-authorizations/execute:
    post:
      operationId: authorize_agent_tool_execute_internal_agent_runs_tool_authorizations_execute_post
      parameters:
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentToolExecuteAuthorizationRequest'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Authorize Agent Tool Execute
      tags:
      - internal-agent-runs
  /internal/agent/runs/tool-authorizations/read:
    post:
      operationId: authorize_agent_tool_read_internal_agent_runs_tool_authorizations_read_post
      parameters:
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentToolReadAuthorizationRequest'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Authorize Agent Tool Read
      tags:
      - internal-agent-runs
  /internal/agent/runs/{run_id}:
    get:
      operationId: get_markdown_agent_run_internal_agent_runs__run_id__get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: query
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      - in: query
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkdownAgentRunStatusResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Markdown Agent Run
      tags:
      - internal-agent-runs
...